+++ DICOM/Network/DicomPresentationContextCollection.cs
@@ -104,25 +104,35 @@ namespace Dicom.Network
         {
             if (request is DicomCStoreRequest cstore)
             {
-                var pcs = _pc.Values.Where(x => x.AbstractSyntax == request.SOPClassUID);
-                if (cstore.TransferSyntax == DicomTransferSyntax.ImplicitVRLittleEndian)
-                {
-                    pcs = pcs.Where(x => x.GetTransferSyntaxes().Contains(DicomTransferSyntax.ImplicitVRLittleEndian));
-                }
-                else
-                {
-                    pcs = pcs.Where(x => x.HasTransferSyntax(cstore.TransferSyntax));
-                }
+                var pcs = _pc.Values
+                    .Where(x => x.AbstractSyntax == request.SOPClassUID)
+                    .Where(x => x.HasTransferSyntax(cstore.TransferSyntax));
 
                 var pc = pcs.FirstOrDefault();
                 if (pc == null)
                 {
-                    var tx = new List<DicomTransferSyntax>();
-                    if (cstore.TransferSyntax != DicomTransferSyntax.ImplicitVRLittleEndian) tx.Add(cstore.TransferSyntax);
-                    if (cstore.AdditionalTransferSyntaxes != null) tx.AddRange(cstore.AdditionalTransferSyntaxes);
-                    tx.Add(DicomTransferSyntax.ImplicitVRLittleEndian);
+                    // add a presentation context for the original transfer syntax
+                    if (cstore.TransferSyntax != null)
+                    {
+                        Add(cstore.SOPClassUID, cstore.TransferSyntax);
+                    }
 
-                    Add(cstore.SOPClassUID, tx.ToArray());
+                    // then add another presentation context for additional transfersyntaxes, if provided by the user,
+                    // and for the mandatory ImplicitVRLittleEndian, if the original file was not implicitLittleEndian
+                    var tx = new List<DicomTransferSyntax>();
+                    if (cstore.AdditionalTransferSyntaxes != null)
+                    {
+                        tx.AddRange(cstore.AdditionalTransferSyntaxes);
+                    }
+                    if (cstore.TransferSyntax != DicomTransferSyntax.ImplicitVRLittleEndian)
+                    {
+                        tx.Add(DicomTransferSyntax.ImplicitVRLittleEndian);
+                    }
+
+                    if (tx.Any())
+                    {
+                        Add(cstore.SOPClassUID, tx.ToArray());
+                    }
                 }
             }
             else