+++ ChangeLog.md
@@ -2,6 +2,7 @@
 * Update to DICOM Standard 2020b.
 * Handle N-Event requests in DicomClient (required for synchronous storage commitment) (#1001)
 * Bug fix: Fix DICOM Tag conversion for private tags for element numbers xxff (#1059)
+* ImplicitVRLittleEndian always added to presentation context by default causes other transfer syntax to not be accepted (#1048)
 
 #### v.4.0.5 (5/18/2020)
 * Bug fix: DicomTags of ValueRepresentation LT have not been validated.